Rating Server Progress
In the Android Manifest, I added some Internet permissions in order to
access the device's Wi-Fi. I'm adding some code that should check to see
if the user has internet connection. Then, it should go to the server
domain that I have yet to set up and retrieve or write the data there.
The code is not complete yet, but I'm working on it. Most of the code is
from the Android Studio tutorial, but I changed a few things to suit our
app.